Many cool discoveries for the ctags plugin

+++ vim-utils/autoload/utils.vim
@@ -430,14 +430,14 @@ function! utils#UpdateCscope() abort
 		return
 	endif
 
-	" if executable('rg') && has('nvim') && has('python3') " Use asynch nvim call instead
-		" call UpdateTagsRemote()
-		" return	
+	if executable('rg') && has('nvim') && has('python3') " Use asynch nvim call instead
+		call UpdateTagsRemote()
+		return	
 	" elseif has('python3')			" If python3 is available use it
 	" if has('python3')			" If python3 is available use it
 		" call python#UpdateCtags()
 		" return
-	" endif
+	endif
 
 	silent! cs kill -1
 	if has('unix')
